Disable SIP by default
SIP is not needed in such an unofficial environment that requires tinkering., in fact it makes modifications impossible to do AND it is very painfully hard to disable after installation is done. This should NOT BE ENABLED by default so I'm proposing a change for it to be disabled. Much like when you're repairing your car, it isn't advisable to have the car locked and armed only for the alarm to go off the minute you touch something. YOu should leave it unlocked, just as SIP. Please set to 77, thanks.